实现logout api并且增加test

+++ puluo_app_backend/core/src/main/java/com/puluo/api/result/UserLogoutResult.java
@@ -5,8 +5,8 @@ import com.puluo.util.HasJSON;
 
 public class UserLogoutResult extends HasJSON {
 	public String uuid;
-	public int duration_seconds;
-	public UserLogoutResult(String uuid, int duration_seconds) {
+	public long duration_seconds;
+	public UserLogoutResult(String uuid, long duration_seconds) {
 		super();
 		this.uuid = uuid;
 		this.duration_seconds = duration_seconds;
